allow multiqueued execution at the Local Executor
Motivation
==========

Threaded work usually partition many classes of tasks by spawning multiple
threads and managing aspects of how they run.

Work that runs on the same CPU don't have that luxury: as new work is generated
it is queued and then drawn from in FIFO order.

One example is a database to which reads and writes are issued: If one issues
1,000,000 writes at once and 1 read, that read now has to sit in queue until
all those 1,000,000 writes complete as there is no other thread to preempt and
jump to. In a threaded system preemption eventually happens, so as long as those
example actors are kept in different threads progress is guaranteed.

This is problematic for systems that depend on local CPU execution, especially
ones that want to guarantee some level of fairness and latency among different
actors that are user-driven with unpredictable patterns.

The solution
============

My solution is to move the queue inside the executor to a Vector. Classes for
now can't be destroyed, so that works well. We may have to use a Set to allow
for classes to be destroyed. Because we need to keep updating that index, and
because we may add, and in the future remove, classes, the queues and index are
kept in an interior-mutability RefCell lumped in a state structure, that is
internal to this module.

The user of this interface can only see an opaque handler, so this will allow us
to use more complex datastructures in the future without changing the API.

Together with this vector we keep a round-robin index of which queue to probe
next.  Semantics are changed a bit so returning false in tick() means that no
queue has any work to do.

Inspiration and future work
===========================

This work is inspired by the Seastar C++ Thread-per-Core framework, that underpins
technologies like ScyllaDB and RedPanda. I view it as a requirement for a good and
reliable implementation of a Thread-per-Core system in Rust.

Each actor (readers, writers, LSM compactors) is added to a different queue. At this
point I am looking for feedback on the direction and the rust-ness of the proposal.
So we just cycle round-robin through the many classes, which already demonstrates some
level of fairness in my example application.

As a next step, should this be well received, I intend to implement a fully fledged
scheduler like the one Seastar has, where the application can configure shares that
the various actors will have. That scheduler is in turn inspired by the Linux CFQ,
and I don't see a reason why we'd have to do anything different here.